Rework browser delegation into host browser crate
