+++ hosts/browser/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "mech-host-browser"
+version = "0.3.5"
+authors = ["Corey Montella <corey@mech-lang.org>"]
+description = "Browser host integration for Mech runtimes."
+documentation = "https://docs.mech-lang.org"
+homepage = "https://mech-lang.org"
+repository = "https://github.com/mech-lang/mech"
+license = "Apache-2.0"
+edition = "2024"
+rust-version = "1.92"
+
+[features]
+default = []
+serde = ["dep:serde", "dep:serde_derive", "mech-core/serde", "mech-runtime/serde"]
+provider = ["mech-core/string"]
+delegation = ["serde", "mech-runtime/host_delegation"]
+delegation_signing = ["delegation", "mech-runtime/host_delegation_signing"]
+
+[dependencies]
+mech-core = { version = "0.3.5", default-features = false }
+mech-runtime = { version = "0.3.5", default-features = false }
+serde = { version = "1.0.228", optional = true }
+serde_derive = { version = "1.0.228", optional = true }
+serde_json = "1.0.150"
+base64 = "0.22.1"
+
+[dev-dependencies]
+serde_json = "1.0.150"
