+++ hosts/browser/src/lib.rs
@@ -1,0 +1,13 @@
+pub mod config;
+#[cfg(feature = "provider")]
+pub mod provider;
+
+#[cfg(feature = "delegation")]
+pub mod delegation;
+
+pub use self::config::*;
+#[cfg(feature = "provider")]
+pub use self::provider::*;
+
+#[cfg(feature = "delegation")]
+pub use self::delegation::*;
