--- src/runtime/src/browser.rs
@@ -1,9 +1,0 @@
-pub use mech_core::{
-    browser_capability_error, browser_capability_result, BrowserAuthority, BrowserBudget,
-    BrowserCapabilityError, BrowserCapabilityGrant, BrowserCapabilityRequest,
-    BrowserDomManifestEntry, BrowserDomPath, BrowserDomProperty, BrowserDomScope,
-    BrowserNetworkScope, BrowserOperation, BrowserResource, BrowserResourceKind,
-    BrowserStorageBackend, BrowserStorageScope, BROWSER_CLIPBOARD_PROVIDER_URI,
-    BROWSER_DOM_PROVIDER_URI, BROWSER_HOST_IDENTITY, BROWSER_NETWORK_PROVIDER_URI,
-    BROWSER_STORAGE_PROVIDER_URI,
-};
