+++ src/runtime/src/lib.rs
@@ -1,55 +1,81 @@
 #![cfg_attr(feature = "no_std", no_std)]
 
-pub mod runtime;
 pub mod id;
-pub mod capability;
 pub mod config;
 pub mod config_profile;
-pub mod browser_host_config;
-pub mod browser;
 #[cfg(feature = "host_delegation")]
 pub mod host_delegation;
 #[cfg(feature = "host_delegation_signing")]
 pub mod host_delegation_crypto;
-mod config_spec;
-pub mod store;
-pub mod resolver;
-pub mod event;
-pub mod context;
 mod resource;
+
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod runtime;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod capability;
+#[cfg(any(feature = "program", feature = "compiler"))]
+mod config_spec;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod store;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod resolver;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod event;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub mod context;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod host;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod scheduler;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod transaction;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod actor;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod service;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod actor_behavior;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub mod module;
-#[cfg(feature = "watcher")]
+#[cfg(all(feature = "watcher", any(feature = "program", feature = "compiler")))]
 mod workspace;
 
 pub use self::id::*;
-pub use self::runtime::*;
 pub use self::config::*;
 pub use self::config_profile::*;
-pub use self::browser_host_config::*;
-pub use self::browser::*;
 #[cfg(feature = "host_delegation")]
 pub use self::host_delegation::*;
 #[cfg(feature = "host_delegation_signing")]
 pub use self::host_delegation_crypto::*;
-pub use self::config_spec::*;
-pub use self::capability::*;
-pub use self::store::*;
-pub use self::resolver::*;
-pub use self::event::*;
-pub use self::context::*;
 pub use self::resource::*;
+
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::runtime::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::config_spec::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::capability::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::store::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::resolver::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::event::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
+pub use self::context::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::host::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::scheduler::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::transaction::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::actor::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::service::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::actor_behavior::*;
+#[cfg(any(feature = "program", feature = "compiler"))]
 pub use self::module::*;
-#[cfg(feature = "watcher")]
+#[cfg(all(feature = "watcher", any(feature = "program", feature = "compiler")))]
 pub use self::workspace::*;
