+++ src/wasm/Cargo.toml
@@ -63,9 +63,9 @@ program = ["mech-core/program", "mech-interpreter/program", "mech-syntax/program
 pretty_print = ["mech-core/pretty_print", "mech-syntax/pretty_print", "mech-runtime/pretty_print"]
 formatter = ["mech-syntax/formatter"]
 mika = ["mech-syntax/mika", "mech-core/mika", "mech-runtime/mika"]
-serde = ["mech-core/serde", "mech-syntax/serde", "mech-interpreter/serde", "mech-runtime/serde"]
-host_delegation = ["serde", "mech-runtime/host_delegation"]
-host_delegation_signing = ["host_delegation", "mech-runtime/host_delegation_signing"]
+serde = ["mech-core/serde", "mech-syntax/serde", "mech-interpreter/serde", "mech-runtime/serde", "mech-host-browser/serde"]
+host_delegation = ["serde", "mech-runtime/host_delegation", "mech-host-browser/delegation"]
+host_delegation_signing = ["host_delegation", "mech-runtime/host_delegation_signing", "mech-host-browser/delegation_signing"]
 run_program = ["program"]
 inline_output_values = []
 codeblock_output_values = []
@@ -294,6 +294,7 @@ mech-core = {version = "0.3.5", default-features = false}
 mech-syntax = { version = "0.3.5", default-features = false}
 mech-interpreter = { version = "0.3.5", default-features = false}
 mech-runtime = { path = "../runtime", default-features = false }
+mech-host-browser = { path = "../../hosts/browser", default-features = false, features = ["provider"] }
 
 wasm-bindgen = "0.2.100"
 serde-wasm-bindgen = "0.6.5"
